+++ chat/fixed-chat-window/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @lg-chat/fixed-chat-window
 
+## 3.0.11
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/popover@13.0.8
+  - @lg-chat/chat-window@3.0.10
+  - @lg-chat/title-bar@3.0.10
+  - @leafygreen-ui/icon@13.2.1
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/tokens@2.12.1
+  - @leafygreen-ui/typography@20.1.6
+
 ## 3.0.10
 
 ### Patch Changes