+++ packages/badge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/badge
 
+## 9.0.7
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/tokens@2.12.1
+
 ## 9.0.6
 
 ### Patch Changes